Created a new module called files.py.
