+++ src/pyspextool/io/__init__.py
@@ -1,8 +1,3 @@
-"""
-This subpackage contains functions for FITS file input/output.
-"""
-
-__all__ = ['average_header_info', 'check_directory', 'check_file', 'extract_filestring',
-           'get_header_info', 'make_fullpath', 'read_flat_fits', 'read_flatcal_file',
-           'read_instrument_file', 'read_uspex_fits', 'read_wavecal_file',
-           'write_flat_fits']
+__all__ = ['average_header_info', 'files.py', 'get_header_info',
+           'read_flat_fits', 'read_flatcal_file', 'read_instrument_file',
+           'read_uspex_fits', 'read_wavecal_file', 'write_flat_fits']
